backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
alembic>=1.7.0
pydantic>=1.8.0
python-jose>=3.3.0
passlib>=1.7.0
bcrypt>=3.2.0
python-multipart>=0.0.5
stripe>=2.60.0
redis>=4.0.0
celery>=5.2.0
gunicorn>=20.1.0
email-validator>=1.1.0
slowapi>=0.1.5